% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_fluorochromes.R
\name{clean_fluorochromes}
\alias{clean_fluorochromes}
\title{Adapt fluorochrome names to preferred short name
from https://github.com/ISAC-DSTF/ProbeTagDictionary/blob/master/ProbeTagDictionary.json}
\usage{
clean_fluorochromes(x)
}
\arguments{
\item{x}{Fluorochromes to clean}
}
\description{
Adapt fluorochrome names to preferred short name
from https://github.com/ISAC-DSTF/ProbeTagDictionary/blob/master/ProbeTagDictionary.json
}
\examples{
clean_fluorochromes(c("Bv-421", "Pacificblue"))

}
