% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_brightness.R
\name{estimate_SI_single}
\alias{estimate_SI_single}
\title{Estimate Stain Index for one detector}
\usage{
estimate_SI_single(
  ff,
  detector,
  cutoff,
  meta = data.frame(matrix(NA, nrow = 1, ncol = 0))
)
}
\arguments{
\item{ff}{FlowFrame. Assumed to be compensated and pregated.}

\item{detector}{Detector to evaluate. Should be a column name of ff.}

\item{cutoff}{Threshold between positive and negative population.}

\item{meta}{Dataframe of 1 row with relevant meta information.
The results will be appended to this.}
}
\description{
Estimate Stain Index for one detector
}
