% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_brightness.R
\name{estimate_spillover}
\alias{estimate_spillover}
\title{Estimate spillover for one detector}
\usage{
estimate_spillover(ff, detector, other_detectors, meta)
}
\arguments{
\item{ff}{FlowFrame. Assumed to be pregated.}

\item{detector}{Detector to evaluate. Should be a column name of ff.}

\item{other_detectors}{Detectors to computer spillover into.}

\item{meta}{Dataframe of 1 row with relevant meta information,
as returned by estimate_SI. Should at least contain
"Cutoff" column.}
}
\description{
Estimate spillover for one detector
}
