% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_brightness.R
\name{estimate_spread}
\alias{estimate_spread}
\title{Estimate spread on a number of single stains}
\usage{
estimate_spread(
  single_stains,
  transform = TRUE,
  unmixing_M = NULL,
  compensation_M = NULL
)
}
\arguments{
\item{single_stains}{Data frame with ID, File, Fluorochrome and Detector columns}

\item{transform}{To be passed onto find_cutoff_flowDensity. If TRUE (default),
estimateLogicle will be applied.}

\item{unmixing_M}{Unmixing matrix. Default NULL. If NULL, either assume
the data is already unmixed, or it is not spectral data}

\item{compensation_M}{Compensation matrix. Default NULL. If NULL, assume
the data is already compensated or no compensation is
necessary.}
}
\description{
Estimate spread on a number of single stains
}
