% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_brightness.R
\name{extract_spillover}
\alias{extract_spillover}
\title{Extract a compensation matrix}
\usage{
extract_spillover(SI, singles_of_interest = seq_len(nrow(SI)))
}
\arguments{
\item{SI}{Result of estimate_brightness with estimate_spillover = TRUE}

\item{singles_of_interest}{Rownames of the singles you want to include.
Default is all.}
}
\description{
Extract a compensation matrix
}
