% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_brightness.R
\name{extract_spread}
\alias{extract_spread}
\title{Extract a SSM matrix}
\usage{
extract_spread(
  SI,
  singles_of_interest = seq_len(nrow(SI)),
  detectors = SI[singles_of_interest, "Detector"]
)
}
\arguments{
\item{SI}{Result of estimate_brightness with estimate_spread = TRUE}

\item{singles_of_interest}{Rownames of the singles you want to include.
Default is all.}
}
\description{
Extract a SSM matrix
}
