% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cutoff.R
\name{find_cutoff_FMO}
\alias{find_cutoff_FMO}
\title{Decide positivity cutoff based on quantile}
\usage{
find_cutoff_FMO(ff, detector, quantile = 0.9999)
}
\arguments{
\item{ff}{FlowFrame to use}

\item{detector}{Column name of the flow frame}

\item{quantile}{Quantile used for the cutoff. Default = 0.9999}
}
\description{
Mainly useful when using an FMO or unstained sample.
}
