% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_plateau.R
\name{find_plateau}
\alias{find_plateau}
\title{Find plateau from numeric vector}
\usage{
find_plateau(data)
}
\arguments{
\item{data}{Vector with values}
}
\value{
index of the start of the plateau
}
\description{
Takes every point as possible split,
fits linear model on left side and horizontal line on right side,
and looks for the point with minimal residuals.
}
