% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_helper_functions.R
\name{get_highest_channel}
\alias{get_highest_channel}
\title{Identify channel with highest value at a certain quantile level}
\usage{
get_highest_channel(ff, q = 0.9, ref_ff = NULL)
}
\arguments{
\item{ff}{FlowFrame}

\item{q}{quantile for which the values should be evaluated}
}
\description{
Identify channel with highest value at a certain quantile level
}
