% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicate_optimal_voltages.R
\name{indicate_optimal_voltages_single}
\alias{indicate_optimal_voltages_single}
\title{Estimate optimal SI for a single voltration}
\usage{
indicate_optimal_voltages_single(
  SI,
  max_outOfRange = 0.01,
  max_n_rSD = 3,
  method = "plateau"
)
}
\arguments{
\item{SI}{Dataframe with at least the following columns:
Pctg_OutOfRange, rSD, SI.
For example as part of the object generated by estimate_brightness.}

\item{max_n_rSD}{Threshold for rSD value is determined by multiplying this
number with the minimum rSD occuring during the voltration.
Only used if method == "rSD threshold"}

\item{method}{"plateau" or "rSD threshold"}
}
\value{
vector with 1 for optimal value, 0 for suboptimal values and NA
         for rows which were not considered due to too many cells out of
         range or too large rSD
}
\description{
Estimate optimal SI for a single voltration
}
