% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_brightness.R
\name{plot_brightness}
\alias{plot_brightness}
\title{Plot SI as computed by estimate_brightness}
\usage{
plot_brightness(
  SI_optimal,
  cells = NULL,
  cofactor = 150,
  order = "SI",
  decreasing = TRUE
)
}
\arguments{
\item{SI_optimal}{SI result to plot}

\item{cells}{As returned by estimate_brightness, to show signal intensity}

\item{cofactor}{asinh cofactor used to plot signal intensity}

\item{order}{Column of SI_optimal used for sorting}

\item{decreasing}{Argument for sorting}
}
\description{
Plot SI as computed by estimate_brightness
}
