% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_voltration.R
\name{plot_voltrations}
\alias{plot_voltrations}
\title{Plot voltrations}
\usage{
plot_voltrations(
  SI,
  optimal_voltage = NULL,
  dummy_voltages = NULL,
  cofactor = 150
)
}
\arguments{
\item{SI}{Stain index list as generated by estimate_brightness}

\item{optimal_voltage}{Rownames of the SI$SI data.frame to use}

\item{dummy_voltages}{Voltages to grey out in the overview}

\item{cofactor}{Arcsinh cofactor used for visualisation}
}
\description{
Plot voltrations
}
