% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_cells.R
\name{sample_cells}
\alias{sample_cells}
\title{Sample cells in long format}
\usage{
sample_cells(ff, detector, meta, n = 6000)
}
\arguments{
\item{ff}{flowFrame}

\item{detector}{Column name of flowFrame}

\item{meta}{Dataframe with extra metadata describing the flowFrame. Will be
added as additional columns to the result}

\item{n}{Number of cells to sample at most. If flowFrame contains fewer cells,
nrow(ff) will be sampled.}
}
\description{
Sample cells in long format
}
