% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_theoretical_brightness.R
\name{theoretical_brightness}
\alias{theoretical_brightness}
\title{Function to load theoretical brightness}
\usage{
theoretical_brightness()
}
\value{
A data frame which has a Fluorochrome and a Brightness column
}
\description{
Function to load theoretical brightness
}
