% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_brightness.R
\name{estimate_brightness}
\alias{estimate_brightness}
\title{Estimate SI brightness from single stains}
\usage{
estimate_brightness(
  single_stains,
  return_cells = TRUE,
  seed = 1,
  comp = NULL,
  transform = TRUE,
  unstained = NULL,
  remove_neg_from_single = FALSE,
  estimate_spillover = FALSE,
  silent = TRUE,
  method = "plateau"
)
}
\arguments{
\item{single_stains}{Dataframe with at least the following columns:
ID, Fluorochrome, Detector, File.
ID is assumed to be unique.
Detector should correspond with the colnames of the
flow frames.
File should be full path to the single stain fcs files,
which have been preprocessed (e.g. clean up gates been
applied).}

\item{return_cells}{Boolean. If true, a list with subsampled cell values
are returned, e.g. for plotting.}

\item{seed}{Seed for reproducability}

\item{comp}{Compensation matrix.}

\item{transform}{Transform parameter to be passed to find_cutoff_flowDensity.
Can either be a logical value or a transformList.
If FALSE, no transform is applied. If TRUE (default),
flowcore::estimateLogicle is called, and if this fails, the
default logicleTransform() is applied. If a transformList,
this transformList is applied. Note that the value is
returned in the original space.}

\item{unstained}{If this parameter is provided, it should contain a path
to a preprocessed unstained fcs file. The positivity
threshold will be defined on this sample, and the cells
will be concatened to the single stain for the calculations.}

\item{method}{"plateau" or "rSD threshold"}
}
\description{
Estimate SI brightness from single stains
}
