% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_brightness.R
\name{estimate_brightness}
\alias{estimate_brightness}
\title{Estimate SI from single stains}
\usage{
estimate_brightness(
  single_stains,
  preprocessing_parameters = list(compensate = FALSE, removeMargins = TRUE,
    removeDoublets = TRUE, pregate = list(list(marker_values = list(`FSC-A` = c(min =
    60000, range_min = 50000, max = 125000, range_max = 50000), `SSC-A` = c(max = 80000,
    range_max = 50000)))), pregate_tf = FALSE),
  return_cells = TRUE,
  seed = 1,
  comp = NULL,
  transform = TRUE,
  unstained = NULL,
  estimate_spillover = FALSE
)
}
\arguments{
\item{single_stains}{Dataframe with at least the following columns:
ID, Fluorochrome, Detector, File.
ID is assumed to be unique.
Detector should correspond with the colnames of the
flow frames.
File should be full path to the single stain fcs files.}

\item{preprocessing_parameters}{Parameters for pregating done by flowDensity.
Named list with booleans (compensate, removeMargins,
removeDoublets) and "pregate", which is a list itself,
and contains per gate a list marker_values.
The name corresponding to the channel.}

\item{return_cells}{Boolean. If true, a list with cell values are returned}

\item{seed}{Seed for reproducability}

\item{comp}{Compensation matrix.}

\item{transform}{Can either be a logical value or a transformList.
If FALSE, no transform is applied. If TRUE (default),
flowcore::estimateLogicle is called, and if this fails, the
default logicleTransform() is applied. If a transformList,
this transformList is applied. Note that the value is
returned in the original space.}

\item{unstained}{If this parameter is provided, it should contain a path
to an fcs file. The unstained population will then be
taken from this file rather than the single stain.}
}
\description{
Estimate SI from single stains
}
