% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_gate.R
\name{estimate_gate}
\alias{estimate_gate}
\title{For pregating of a flowFrame}
\usage{
estimate_gate(
  ff,
  marker_values = list(`FSC-A` = c(min = 60000, max = 125000, range_min = 50000,
    range_max = 50000), `SSC-A` = c(max = 80000, range_max = 50000)),
  plot = FALSE,
  plot_title = "Gate",
  tinypeak.removal = 1/25
)
}
\arguments{
\item{ff}{flowFrame to gate}

\item{marker_values}{List with a vector for every marker to be gated,
which can contain an estimated minimum threshold (min)
and the acceptable range to flowDensity cutoff can be in
(from min - range_min to min + range_min). Same for max
and max_range. If no possible cut is found in this range,
the default value is returned (as passed in min or max).}

\item{plot}{Boolean whether to plot}

\item{plot_title}{Title for the plot. Default = "Gate"}

\item{tinypeak.removal}{Passed to flowDensity::deGate, default 1/25}
}
\description{
For pregating of a flowFrame
}
