% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicate_optimal_voltages.R
\name{indicate_optimal_voltages}
\alias{indicate_optimal_voltages}
\title{Estimate optimal SI for a single voltration}
\usage{
indicate_optimal_voltages(SI)
}
\arguments{
\item{SI}{Dataframe with at least the following columns:
Group, Pctg_OutOfRange, rSD, SI.
For example as part of the object generated by estimate_brightness.}
}
\value{
The same object, in which an additional column "Optimal"
         is added in the SI dataframe, which will have one TRUE value per
         group
}
\description{
Estimate optimal SI for a single voltration
}
