% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_FCS.R
\name{prep_FCS}
\alias{prep_FCS}
\title{Preprocess FCS file}
\usage{
prep_FCS(
  file,
  compensate = "SPILL",
  removeMargins = TRUE,
  removeDoublets = TRUE,
  pregate = TRUE,
  pregate_tf = TRUE,
  ...
)
}
\arguments{
\item{file}{Path to fcs file or flowFrame.
Will be read with truncate_max_range = FALSE. Additional parameters
for read.FCS will be passed along.}

\item{compensate}{Can be FALSE (no compensation), a character (keyword in
the flowFrame containing the compensation matrix) or a
numeric matrix (compensation matrix to use).
Default = "SPILL"}

\item{removeMargins}{Can be FALSE (no margin removal), TRUE (margin removal
for all channels except those containing File or
Original_ID) or a character vector (containing the
names of the channels to remove the margins for).
Default = TRUE}

\item{removeDoublets}{Boolean, whether or not to remove}

\item{pregate}{Pregating information}

\item{pregate_tf}{TransformList to use in pregating.}

\item{...}{Extra arguments to pass to read.FCS}
}
\description{
Preprocess FCS file
}
